# services/api/internal/interface/http/gen/openapi.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package gen

import (
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// ListApplicationsParams defines parameters for ListApplications.
type ListApplicationsParams struct {
	Status *string `form:"status,omitempty" json:"status,omitempty"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /applications)
	ListApplications(ctx echo.Context, params ListApplicationsParams) error

	// (POST /applications)
	CreateApplication(ctx echo.Context) error

	// (GET /applications/{id})
	GetApplication(ctx echo.Context, id string) error

	// (PATCH /applications/{id}/next-action)
	UpdateNextAction(ctx echo.Context, id string) error

	// (PATCH /applications/{id}/status)
	UpdateStatus(ctx echo.Context, id string) error

	// (POST /checklist-items)
	CreateChecklistItem(ctx echo.Context) error

	// (PATCH /checklist-items/{id})
	UpdateChecklistItem(ctx echo.Context, id string) error

	// (GET /companies)
	ListCompanies(ctx echo.Context) error

	// (POST /companies)
	CreateCompany(ctx echo.Context) error

	// (GET /dashboard)
	GetDashboard(ctx echo.Context) error

	// (POST /events)
	CreateEvent(ctx echo.Context) error

	// (GET /events/upcoming)
	GetUpcomingEvents(ctx echo.Context) error

	// (GET /health)
	GetHealth(ctx echo.Context) error

	// (POST /interviews)
	CreateInterview(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// ListApplications converts echo context to params.
func (w *ServerInterfaceWrapper) ListApplications(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params ListApplicationsParams
	// ------------- Optional query parameter "status" -------------

	err = runtime.BindQueryParameter("form", true, false, "status", ctx.QueryParams(), &params.Status)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter status: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ListApplications(ctx, params)
	return err
}

// CreateApplication converts echo context to params.
func (w *ServerInterfaceWrapper) CreateApplication(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateApplication(ctx)
	return err
}

// GetApplication converts echo context to params.
func (w *ServerInterfaceWrapper) GetApplication(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, ctx.Param("id"), &id)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetApplication(ctx, id)
	return err
}

// UpdateNextAction converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateNextAction(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, ctx.Param("id"), &id)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UpdateNextAction(ctx, id)
	return err
}

// UpdateStatus converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateStatus(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, ctx.Param("id"), &id)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UpdateStatus(ctx, id)
	return err
}

// CreateChecklistItem converts echo context to params.
func (w *ServerInterfaceWrapper) CreateChecklistItem(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateChecklistItem(ctx)
	return err
}

// UpdateChecklistItem converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateChecklistItem(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, ctx.Param("id"), &id)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UpdateChecklistItem(ctx, id)
	return err
}

// ListCompanies converts echo context to params.
func (w *ServerInterfaceWrapper) ListCompanies(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ListCompanies(ctx)
	return err
}

// CreateCompany converts echo context to params.
func (w *ServerInterfaceWrapper) CreateCompany(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateCompany(ctx)
	return err
}

// GetDashboard converts echo context to params.
func (w *ServerInterfaceWrapper) GetDashboard(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetDashboard(ctx)
	return err
}

// CreateEvent converts echo context to params.
func (w *ServerInterfaceWrapper) CreateEvent(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateEvent(ctx)
	return err
}

// GetUpcomingEvents converts echo context to params.
func (w *ServerInterfaceWrapper) GetUpcomingEvents(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetUpcomingEvents(ctx)
	return err
}

// GetHealth converts echo context to params.
func (w *ServerInterfaceWrapper) GetHealth(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetHealth(ctx)
	return err
}

// CreateInterview converts echo context to params.
func (w *ServerInterfaceWrapper) CreateInterview(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateInterview(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/applications", wrapper.ListApplications)
	router.POST(baseURL+"/applications", wrapper.CreateApplication)
	router.GET(baseURL+"/applications/:id", wrapper.GetApplication)
	router.PATCH(baseURL+"/applications/:id/next-action", wrapper.UpdateNextAction)
	router.PATCH(baseURL+"/applications/:id/status", wrapper.UpdateStatus)
	router.POST(baseURL+"/checklist-items", wrapper.CreateChecklistItem)
	router.PATCH(baseURL+"/checklist-items/:id", wrapper.UpdateChecklistItem)
	router.GET(baseURL+"/companies", wrapper.ListCompanies)
	router.POST(baseURL+"/companies", wrapper.CreateCompany)
	router.GET(baseURL+"/dashboard", wrapper.GetDashboard)
	router.POST(baseURL+"/events", wrapper.CreateEvent)
	router.GET(baseURL+"/events/upcoming", wrapper.GetUpcomingEvents)
	router.GET(baseURL+"/health", wrapper.GetHealth)
	router.POST(baseURL+"/interviews", wrapper.CreateInterview)

}