# services/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Shukatsu Flow API
  version: 0.1.0

servers:
  - url: http://localhost:8080

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]

    HealthResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]

    Company:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        industry:
          type: string
          nullable: true
        job_type:
          type: string
          nullable: true
        preference_level:
          type: integer
        memo:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name]

    Application:
      type: object
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        application_title:
          type: string
        status:
          type: string
        next_action_text:
          type: string
          nullable: true
        next_action_at:
          type: string
          format: date-time
          nullable: true
      required: [id, company_id, status]

    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        application_id:
          type: string
        kind:
          type: string
        title:
          type: string
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
          nullable: true
      required: [id, application_id, kind, title]

    Interview:
      type: object
      properties:
        id:
          type: string
        event_id:
          type: string
        round:
          type: integer
        interview_type_id:
          type: string
        memo:
          type: string
      required: [id, event_id, round]

    ChecklistItem:
      type: object
      properties:
        id:
          type: string
        application_id:
          type: string
        title:
          type: string
        is_done:
          type: boolean
      required: [id, application_id, title]

paths:
  /health:
    get:
      operationId: getHealth
      responses:
        "200":
          description: OK

  /companies:
    get:
      operationId: listCompanies
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK

    post:
      operationId: createCompany
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created

  /applications:
    get:
      operationId: listApplications
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: OK

    post:
      operationId: createApplication
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created

  /applications/{id}:
    get:
      operationId: getApplication
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK

  /applications/{id}/status:
    patch:
      operationId: updateStatus
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated

  /applications/{id}/next-action:
    patch:
      operationId: updateNextAction
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated

  /dashboard:
    get:
      operationId: getDashboard
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK

  /events:
    post:
      operationId: createEvent
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created

  /events/upcoming:
    get:
      operationId: getUpcomingEvents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK

  /interviews:
    post:
      operationId: createInterview
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created

  /checklist-items:
    post:
      operationId: createChecklistItem
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created

  /checklist-items/{id}:
    patch:
      operationId: updateChecklistItem
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated